Remove target 'pre' from Makefile.pdlibbuilder
Phony target 'pre' forces rebuilds. This should not be default
behavior of Makefile.pdlibbuilder. If one needs a pre-build target
(phony or real file), declare it as prerequisite of $(all.objects)
in the library makefile after inclusion of Makefile.pdlibbuilder.
A pre-build target can have any name that isn't a target in
Makefile.pdlibbuilder.