Fix broken complex initializers (use ISO C99 CMPLX macro).
Under gcc (including mingw), the previous MSVC-style code would cause
the imaginary parts to be ignored, yielding the wrong complex values.

The new code should hopefully work across all supported platforms. On
Windows, which seems to lack the CMPLX macro in its complex.h header, we
define it in a suitable manner depending on the compiler.